move cache code to it's own class
